ReadingsRestructure

+++ app/src/main/java/br/ecosynergy_app/teams/ActiviesData.kt
@@ -1,11 +1,11 @@
 package br.ecosynergy_app.teams
 
     data class ActivitiesRequest(
-        val id : Long
+        val id : Int
     )
 
     data class ActivitiesResponse(
-        val id : Long,
+        val id : Int,
         val name: String,
         val sector: String
     )
